+++ OutOfSchool/OutOfSchool.WebApi/Extensions/Startup/AverageRatingExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using OutOfSchool.BackgroundJobs.Config;
+using OutOfSchool.BackgroundJobs.Jobs;
 using OutOfSchool.BusinessLogic.Common.QuartzConstants;
 using OutOfSchool.BusinessLogic.Services.AverageRatings;
 using Quartz;
