+++ OutOfSchool/OutOfSchool.WebApi/Extensions/Startup/FileStorageExtensions.cs
@@ -1,6 +1,7 @@
 using Google.Cloud.Storage.V1;
 using Microsoft.Extensions.Options;
 using OutOfSchool.BackgroundJobs.Config;
+using OutOfSchool.BackgroundJobs.Jobs;
 using OutOfSchool.BusinessLogic.Common.QuartzConstants;
 using OutOfSchool.BusinessLogic.Services.Gcp;
 using OutOfSchool.BusinessLogic.Util.FakeImplementations;
